backend/requirements/base.txt
fastapi[standard]
pandas==2.2.3
pyarrow==18.1.0
yfinance==0.2.65
SQLAlchemy>=2.0
psycopg2-binary>=2.9
curl_cffi
passlib[bcrypt]>=1.7.4
python-jose[cryptography]